Extracts ConnectIntegration
